requirements.txt
inotify==0.2.10
lxml==5.3.0
nose==1.3.7
psycopg2==2.9.9
PyYAML==6.0.2
